requirements.txt
fastapi==0.70.0
uvicorn==0.15.0

PyJWT==1.7.1 
python-decouple==3.3